requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.4
attrs==24.2.0
auto-py-to-exe==2.44.2
awsebcli==3.21.0
bcrypt==4.2.1
blinker==1.8.2
botocore==1.35.64
bottle==0.13.1
bottle-websocket==0.2.9
cement==2.10.14
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepdiff==8.0.1
dipy==1.9.0
Eel==0.17.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
h5py==3.12.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.0
multidict==6.1.0
mysql-connector-python==9.0.0
mysqlclient==2.2.6
nibabel==5.3.2
numpy==2.1.3
orderly-set==5.2.2
outcome==1.3.0.post0
packaging==24.1
paho-mqtt==2.1.0
pathspec==0.10.1
pefile==2024.8.26
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.9.0
pyparsing==3.1.4
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
selenium==4.25.0
semantic-version==2.10.0
setuptools-scm==8.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stripe==11.1.0
termcolor==2.5.0
TIME-python==0.0.17
tqdm==4.67.0
trio==0.26.2
trio-websocket==0.11.1
trx-python==0.3
twilio==9.3.3
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.4
whichcraft==0.6.1
wsproto==1.2.0
yarl==1.13.1
zope.event==5.0
zope.interface==7.0.3
